test(repl): add config_test.go for NewClientDefaultConfig
ClientConfig and NewClientDefaultConfig moved to config.go. There was
no existing unit test for either — both are exercised indirectly
through every integration test that constructs a client, which is fine
for catching breakage but not for catching a silent default-value
change.

Add two small unit tests:
- TestNewClientDefaultConfig pins every documented default
  (Concurrency=4, TargetBatchTime=DefaultTargetBatchTime, Logger
  non-nil, ServerID >= 1001, and the deliberately-zero fields that
  NewClient translates downstream).
- TestNewClientDefaultConfigServerIDIsFresh pins that two calls
  produce different ServerIDs — a constant ServerID would cause MySQL
  to disconnect concurrent test clients sharing the id.
